Add possibility to scrape Fussballdaten.de
Started to scrape the fussballdaten.de website because there are a lot of restrictions when using ElenaSportAPI (only data for seasons back to 2012, additionally only 80 requests per day). Scraped all seasons for the bundesliga with the results.

There are some current problems with the data and the functions which scrape the data.
1. One matchday consists of possibly multiple days where matches take place (mostly on friday, saturday and sunday). We need to be able to calculate how many matches take place on each day.
2. The get_results_for_season function is not able right now to scrape data for future matches because I extract the results which are obviously not available for future matches. I need to write another function to do that or adjust the current one.