Setup the shiny App
Added a shiny app with a general 3-file setup. server.R, ui.R and global.R which control the App. The global file loads all packages and functions needed using the shiny app. The server and ui files manage the sub-serers and sub-uis. We have a dashboardPage-layout with a sidebar which contains several menuItems. Each sub-server controls a menuItem or if this menuItem contains multiple subItems, there is a sub-server for every subItem. Same goes for the ui.

We also adjusted the Setup.R file to load in all files we need for the App