Add functionality to shiny app
Added multiple elements to the shiny app including a table for matchday results and a table showing the last matches and the next match for a selected team in a selected season.

Additionally, created a function to create unique club names (a mapping table) to be able to join with these names.

Also added functions to retrieve the lineups and fixture stats from the Football API.