New functionality for the API-football API
New functions for extracting the current standing in a season, a function for team_stats in a given league and season. Also a function for getting all events happening during a match (i.e., cards, goals, etc.), a function for player stats (shots, offsides, etc.) and a helper function to extract the data from the API in an appropriate format.

Adjusted the api_football_fixtures_general_complete_check function to also work with the new functions